♻️ refactor: refactor folder structure
